chore: change messages

+++ scripts/programs/miniconda.sh
@@ -1,4 +1,11 @@
-mkdir -p ~/miniconda3
-wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O ~/miniconda3/miniconda.sh
-bash ~/miniconda3/miniconda.sh -b -u -p ~/miniconda3
-rm ~/miniconda3/miniconda.sh
+#!/bin/bash
+
+if [ ! -d ~/miniconda3 ]; then
+    echo "Installing miniconda3..."
+	mkdir -p ~/miniconda3
+	wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O ~/miniconda3/miniconda.sh
+	bash ~/miniconda3/miniconda.sh -b -u -p ~/miniconda3
+	rm ~/miniconda3/miniconda.sh
+else
+	echo "Already installed: miniconda3"
+fi
